MAJOR OVERHAUL
Fixed fonts and images from relative to absolute paths. Condensed different HOC components to just one; when fetching data about blog posts the URL path now reflects the backend, so it is passed directly to the HTTP GET calls. If data is returned, it is displayed, whether a single post or a listing. If no data is returned, display the 404 message.

Renamed some files and cleaned house of all excess, unused and experimental files, including unused fonts. The header was also redone and now displays static and dynamic links; dynamics are categories of blog posts.