REFINING THE FLOWS
Removed more unnecessary files. There is now a file for ROUTES, additional ones will be added later. Content Actions are now passed to withContent HoC through the router, so the component is entirely agnostic about what data it is requesting. Also added a 404 error page, which is rendered if the actions do not return any data (404 error)